WIP: Make filtering work properly for HTTP repositories

+++ src/RoadCaptain.Adapters/LocalDirectoryRouteRepository.cs
@@ -36,17 +36,16 @@ namespace RoadCaptain.Adapters
             return Task.FromResult(Directory.Exists(_settings.Directory));
         }
 
-        public async Task<RouteModel[]> SearchAsync(
-            string? world = null,
+        public async Task<RouteModel[]> SearchAsync(string? world = null,
             string? creator = null,
             string? name = null,
             string? zwiftRouteName = null,
-            decimal? minDistance = null,
-            decimal? maxDistance = null,
-            decimal? minAscent = null,
-            decimal? maxAscent = null,
-            decimal? minDescent = null,
-            decimal? maxDescent = null,
+            int? minDistance = null,
+            int? maxDistance = null,
+            int? minAscent = null,
+            int? maxAscent = null,
+            int? minDescent = null,
+            int? maxDescent = null,
             bool? isLoop = null,
             string[]? komSegments = null,
             string[]? sprintSegments = null)
