+++ src/RoadCaptain.App.Runner/ViewModels/DesignTimeSelectRouteWindowViewModel.cs
@@ -44,9 +44,10 @@ namespace RoadCaptain.App.Runner.ViewModels
             return Task.FromResult(true);
         }
 
-        public Task<RouteModel[]> SearchAsync(string? world = null, string? creator = null, string? name = null, string? zwiftRouteName = null,
-            decimal? minDistance = null, decimal? maxDistance = null, decimal? minAscent = null, decimal? maxAscent = null,
-            decimal? minDescent = null, decimal? maxDescent = null, bool? isLoop = null, string[]? komSegments = null,
+        public Task<RouteModel[]> SearchAsync(string? world = null, string? creator = null, string? name = null,
+            string? zwiftRouteName = null,
+            int? minDistance = null, int? maxDistance = null, int? minAscent = null, int? maxAscent = null,
+            int? minDescent = null, int? maxDescent = null, bool? isLoop = null, string[]? komSegments = null,
             string[]? sprintSegments = null)
         {
             return Task.FromResult(new[]
