+++ src/RoadCaptain/Commands/SearchRouteCommand.cs
@@ -7,12 +7,12 @@ namespace RoadCaptain.Commands
         public string? Creator { get; }
         public string? Name { get; }
         public string? ZwiftRouteName { get; }
-        public decimal? MinDistance { get; }
-        public decimal? MaxDistance { get; }
-        public decimal? MinAscent { get; }
-        public decimal? MaxAscent { get; }
-        public decimal? MinDescent { get; }
-        public decimal? MaxDescent { get; }
+        public int? MinDistance { get; }
+        public int? MaxDistance { get; }
+        public int? MinAscent { get; }
+        public int? MaxAscent { get; }
+        public int? MinDescent { get; }
+        public int? MaxDescent { get; }
         public bool? IsLoop { get; }
         public string[]? KomSegments { get; }
         public string[]? SprintSegments { get; }
@@ -22,12 +22,12 @@ namespace RoadCaptain.Commands
             string? creator = null,
             string? name = null,
             string? zwiftRouteName = null,
-            decimal? minDistance = null,
-            decimal? maxDistance = null,
-            decimal? minAscent = null,
-            decimal? maxAscent = null,
-            decimal? minDescent = null,
-            decimal? maxDescent = null,
+            int? minDistance = null,
+            int? maxDistance = null,
+            int? minAscent = null,
+            int? maxAscent = null,
+            int? minDescent = null,
+            int? maxDescent = null,
             bool? isLoop = null,
             string[]? komSegments = null,
             string[]? sprintSegments = null)
