+++ src/RoadCaptain/Ports/IRouteRepository.cs
@@ -5,17 +5,16 @@ namespace RoadCaptain.Ports
     public interface IRouteRepository
     {
         Task<bool> IsAvailableAsync();
-        Task<RouteModel[]> SearchAsync(
-            string? world = null,
+        Task<RouteModel[]> SearchAsync(string? world = null,
             string? creator = null,
             string? name = null,
             string? zwiftRouteName = null,
-            decimal? minDistance = null,
-            decimal? maxDistance = null,
-            decimal? minAscent = null,
-            decimal? maxAscent = null,
-            decimal? minDescent = null,
-            decimal? maxDescent = null,
+            int? minDistance = null,
+            int? maxDistance = null,
+            int? minAscent = null,
+            int? maxAscent = null,
+            int? minDescent = null,
+            int? maxDescent = null,
             bool? isLoop = null,
             string[]? komSegments = null,
             string[]? sprintSegments = null);
